IssueServise쪽. findIssueById Integer 로 수정.
findProjectById(long id) 도 수정

+++ web-server/src/main/java/com/se14/service/IssueService.java
@@ -12,5 +12,5 @@ public interface IssueService {
     void updateIssueStatus(Project project, User updater, Issue issue, IssueStatus status);
     void addComment(Project project, User commenter, Issue issue, String commentTitle, String commentText);
 
-    Issue findIssueById(Long id);
+    Issue findIssueById(Integer id);
 }
